feat(core): actualiza versión a 0.6.0 y refactoriza facturación
Este commit introduce varios cambios significativos:

1.  **Refactorización del Módulo de Facturación:**
    - Se migra el modelo  de Kotlin a Java para mantener consistencia en el codebase.
    - Se elimina por completo la funcionalidad de  (, , ), simplificando la lógica de persistencia.
    - Se introduce  para centralizar la creación de facturas, mejorando la cohesión y reduciendo la duplicación de código en .

2.  **Actualización de Versión y CI:**
    - La versión del proyecto se incrementa de  a .
    - Se actualiza la versión de JDK a 24 en el workflow de CI () para alinear el entorno de construcción con las nuevas capacidades del lenguaje.

BREAKING CHANGE: La funcionalidad y los endpoints relacionados con  han sido eliminados. La migración de  de Kotlin a Java puede requerir ajustes en los servicios que lo consumen.
