+++ pom.xml
@@ -11,7 +11,7 @@
     </parent>
     <groupId>com.termascacheuta</groupId>
     <artifactId>eterea.core.service</artifactId>
-    <version>0.5.1</version>
+    <version>0.6.0</version>
     <name>eterea.core-service</name>
     <description>API - Eterea - Core</description>
     <properties>
