+++ src/main/java/eterea/core/service/service/extern/FacturacionElectronicaService.java
@@ -3,10 +3,7 @@ package eterea.core.service.service.extern;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.json.JsonMapper;
 import eterea.core.service.client.afip.FacturacionAfipClient;
-import eterea.core.service.model.Snapshot;
-import eterea.core.service.model.Track;
 import eterea.core.service.model.dto.FacturacionDto;
-import eterea.core.service.service.SnapshotService;
 import eterea.core.service.service.TrackService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -18,60 +15,21 @@ import java.util.UUID;
 public class FacturacionElectronicaService {
 
     private final FacturacionAfipClient facturacionAfipClient;
-    private final SnapshotService snapshotService;
     private final TrackService trackService;
 
-    public FacturacionElectronicaService(FacturacionAfipClient facturacionAfipClient, SnapshotService snapshotService, TrackService trackService) {
+    public FacturacionElectronicaService(FacturacionAfipClient facturacionAfipClient,
+                                         TrackService trackService
+    ) {
         this.facturacionAfipClient = facturacionAfipClient;
-        this.snapshotService = snapshotService;
         this.trackService = trackService;
     }
 
-    public FacturacionDto makeFactura(FacturacionDto facturacionDto, Track track) {
+    public FacturacionDto makeFactura(FacturacionDto facturacionDto) {
         log.debug("Processing FactuacionElectronicaService.makeFactura");
-        var json = logFacturacion(facturacionDto);
-        if (track == null) {
-            track = trackService.startTracking("make-factura");
-        }
-        var snapshot = Snapshot.builder()
-                .uuid(UUID.randomUUID().toString())
-                .descripcion("make-factura-facturacion-pre")
-                .payload(json)
-                .trackUuid(track.getUuid())
-                .previousUuid(null)
-                .build();
-        logSnapshot(snapshot);
-        snapshot = snapshotService.add(snapshot);
-        logSnapshot(snapshot);
+        logFacturacion(facturacionDto);
         facturacionDto = facturacionAfipClient.facturador(facturacionDto);
-        json = logFacturacion(facturacionDto);
-        snapshot = Snapshot.builder()
-                .uuid(UUID.randomUUID().toString())
-                .descripcion("make-factura-facturacion-post")
-                .payload(json)
-                .trackUuid(track.getUuid())
-                .previousUuid(snapshot.getUuid())
-                .build();
-        logSnapshot(snapshot);
-        snapshot = snapshotService.add(snapshot);
-        logSnapshot(snapshot);
+        logFacturacion(facturacionDto);
         return facturacionDto;
-    }
-
-    private String logSnapshot(Snapshot snapshot) {
-        try {
-            var json = JsonMapper
-                    .builder()
-                    .findAndAddModules()
-                    .build()
-                    .writerWithDefaultPrettyPrinter()
-                    .writeValueAsString(snapshot);
-            log.debug("Snapshot={}", json);
-            return json;
-        } catch (JsonProcessingException e) {
-            log.debug("Snapshot jsonify error {}", e.getMessage());
-            return null;
-        }
     }
 
     private String logFacturacion(FacturacionDto facturacionDTO) {
